Move CheckCombLoops from passes/ to transforms/
